requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0,<1.27  # ChEX expects numpy<=1.26
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.62.0
matplotlib>=3.5.0
Pillow>=9.0.0
torchmetrics>=0.11.0

# For Label-Free CBM
open_clip_torch>=2.20.0  # BiomedCLIP

# For VLG-CBM (ChEX grounding)
omegaconf>=2.3.0
hydra-core>=1.3.0
pytorch-lightning>=2.0.0
einops>=0.6.0
timm>=0.9.0
transformers>=4.37.0
tokenizers==0.15.0

# Optional: Weights & Biases logging
wandb>=0.15.0

# Additional ChEX preprocessing / utility dependencies
pydicom
scikit-image
scipy
opencv-python
stanza
ftfy
regex
albumentations==1.3.0
ensemble_boxes
lovely_tensors
pycocoevalcap
git+https://github.com/bes-dev/mean_average_precision.git
click
deepdiff
kaggle
sentencepiece
